+++ schema/v1/reference_test.go
@@ -7,7 +7,10 @@ import (
 
 	"github.com/santhosh-tekuri/jsonschema/v5"
 	"github.com/stretchr/testify/require"
+	"google.golang.org/protobuf/types/known/structpb"
 	"gopkg.in/yaml.v3"
+
+	"gitlab.com/gitlab-org/step-runner/proto"
 )
 
 func TestReferenceCustomMethods(t *testing.T) {
@@ -161,4 +164,23 @@ func TestReferenceValidation(t *testing.T) {
 			require.Contains(t, err.Error(), test.wantErr)
 		})
 	}
+}
+
+func TestCompileOCI(t *testing.T) {
+	t.Run("compiles OCI reference", func(t *testing.T) {
+		ref := &Reference{OCI: NewOCIReference("registry.gitlab.com", "project/my-repository", "latest")}
+
+		stepRef, err := ref.compile("my_step", map[string]*structpb.Value{}, map[string]string{})
+		require.NoError(t, err)
+		require.Equal(t, proto.StepReferenceProtocol_spec_def, stepRef.Protocol)
+
+		steps := stepRef.SpecDef.Definition.Steps
+		require.Len(t, steps, 2)
+		require.Equal(t, []string{"oci", "fetch"}, steps[0].Step.Path)
+		require.Equal(t, "registry.gitlab.com", steps[0].Inputs["registry"].GetStringValue())
+		require.Equal(t, "project/my-repository", steps[0].Inputs["repository"].GetStringValue())
+		require.Equal(t, "latest", steps[0].Inputs["tag"].GetStringValue())
+		require.Equal(t, "my_step", steps[1].Name)
+		require.Equal(t, "${{steps.fetch_step_my_step.outputs.fetched_step_path}}", steps[1].Step.Url)
+	})
 }